Feat: create some tests - login and POST & GET users

+++ src/entities/transactionCategories.entity.ts
@@ -2,7 +2,7 @@ import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
 
 @Entity()
 export default class TransactionCategories {
-  @PrimaryGeneratedColumn("increment", { type: "bigint" })
+  @PrimaryGeneratedColumn("increment", { type: "integer" })
   id: number;
 
   @Column({ length: 2 })
